+++ source/formats/gp7/parser.h
@@ -310,6 +310,7 @@ struct Note
     {
         Natural,
         Artificial,
+        Pinch,
         Tap,
         Semi,
         Feedback